Remove SocketProvider
