+++ rpc/clients/typescript/src/websocket_provider.ts
@@ -1,6 +1,10 @@
+import * as EventEmitter from 'eventemitter3';
 import * as WebSocket from 'websocket';
 
-import { SocketProvider } from './socket_provider';
+const READY = 'ready';
+const CONNECT = 'connect';
+const ERROR = 'error';
+const CLOSE = 'close';
 
 const SOCKET_MESSAGE = 'socket_message';
 const SOCKET_READY = 'socket_ready';
@@ -8,11 +12,39 @@ const SOCKET_CLOSE = 'socket_close';
 const SOCKET_ERROR = 'socket_error';
 const SOCKET_CONNECT = 'socket_connect';
 
-export class WebSocketProvider extends SocketProvider {
+export class WebSocketProvider extends EventEmitter {
+    public connection: any;
+    public timeoutIfExists: number|undefined;
+    public subscriptions: any;
     private _host: string;
     private _reconnectionTimeoutMs: number;
+    private _messageId: number;
+    private static _validateJSONRPCResponse(response: any): boolean|Error {
+        if (typeof response === 'object') {
+            if (response.error) {
+                if (response.error instanceof Error) {
+                    return new Error(`Node error: ${response.error.message}`);
+                }
+
+                return new Error(`Node error: ${JSON.stringify(response.error)}`);
+            }
+
+            if (response.result === undefined) {
+                return new Error('Validation error: Undefined JSON-RPC result');
+            }
+
+            return true;
+        }
+
+        return new Error('Validation error: Response should be of type Object');
+    }
     constructor(connection: any, reconnectionTimeout: number = 5000, timeout: number|undefined) {
-        super(connection, timeout);
+        super();
+        this._messageId = 1;
+        this.connection = connection;
+        this.timeoutIfExists = timeout;
+        this.subscriptions = {};
+        this.registerEventListeners();
         // HACK(fabio): @types/websocket out-of-date
         this._host = this.connection.url;
         this._reconnectionTimeoutMs = reconnectionTimeout;
@@ -22,7 +54,25 @@ export class WebSocketProvider extends SocketProvider {
      * @param messageEvent message event
      */
     public onMessage(messageEvent: MessageEvent): void {
-        super.onMessage(messageEvent.data);
+        const response = messageEvent.data;
+        let event;
+
+        let responseObject = response as any;
+        if (typeof response !== 'object') {
+            responseObject = JSON.parse(response);
+        }
+
+        if (Array.isArray(responseObject)) {
+            event = responseObject[0].id;
+        } else if (typeof responseObject.id === 'undefined') {
+            event = this._getSubscriptionEvent(responseObject.params.subscription);
+            responseObject = responseObject.params;
+        } else {
+            event = responseObject.id;
+        }
+
+        (this as EventEmitter).emit(SOCKET_MESSAGE, responseObject);
+        (this as EventEmitter).emit(event, responseObject);
     }
     /**
      * This is the listener for the 'error' event of the current socket connection.
@@ -34,7 +84,9 @@ export class WebSocketProvider extends SocketProvider {
 
             return;
         }
-        super.onError(event);
+        (this as EventEmitter).emit(ERROR, event);
+        (this as EventEmitter).emit(SOCKET_ERROR, event);
+        this.removeAllSocketListeners();
     }
     /**
      * This is the listener for the 'close' event of the current socket connection.
@@ -47,7 +99,61 @@ export class WebSocketProvider extends SocketProvider {
 
             return;
         }
-        super.onClose();
+        (this as EventEmitter).emit(CLOSE, closeEvent);
+        (this as EventEmitter).emit(SOCKET_CLOSE, closeEvent);
+        this.removeAllSocketListeners();
+        (this as EventEmitter).removeAllListeners();
+    }
+    /**
+     * Emits the ready event when the connection is established
+     * @param event Event to emit on ready
+     */
+    public onReady(event: any): void {
+        (this as EventEmitter).emit(READY, event);
+        (this as EventEmitter).emit(SOCKET_READY, event);
+    }
+    /**
+     * Emits the connect event and checks if there are subscriptions defined that should be resubscribed.
+     */
+    public async onConnectAsync(): Promise<void> {
+        const subscriptionKeys = Object.keys(this.subscriptions);
+
+        if (subscriptionKeys.length > 0) {
+            let subscriptionId;
+
+            for (const key of subscriptionKeys) {
+                subscriptionId = await this.subscribeAsync(
+                    this.subscriptions[key].subscribeMethod,
+                    this.subscriptions[key].parameters[0],
+                    this.subscriptions[key].parameters.slice(1),
+                );
+
+                if (key !== subscriptionId) {
+                    delete this.subscriptions[subscriptionId];
+                }
+
+                this.subscriptions[key].id = subscriptionId;
+            }
+        }
+
+        (this as EventEmitter).emit(SOCKET_CONNECT);
+        (this as EventEmitter).emit(CONNECT);
+    }
+    /**
+     * Creates the JSON-RPC payload and sends it to the node.
+     * @param method JSON-RPC method to call
+     * @param parameters parameters to send to method call
+     * @returns response to JSON-RPC call
+     */
+    public async sendAsync(method: string, parameters: any): Promise<any> {
+        const response = await this.sendPayloadAsync(this._toPayload(method, parameters));
+        const validationResult = WebSocketProvider._validateJSONRPCResponse(response);
+
+        if (validationResult instanceof Error) {
+            throw validationResult;
+        }
+
+        return response.result;
     }
     /**
      * Removes the listeners and reconnects to the socket.
@@ -78,6 +184,16 @@ export class WebSocketProvider extends SocketProvider {
                 this.emit('reconnected');
             });
         }, this._reconnectionTimeoutMs);
+    }
+    /**
+     * Removes all socket listeners
+     */
+    public removeAllSocketListeners(): void {
+        (this as EventEmitter).removeAllListeners(SOCKET_MESSAGE);
+        (this as EventEmitter).removeAllListeners(SOCKET_READY);
+        (this as EventEmitter).removeAllListeners(SOCKET_CLOSE);
+        (this as EventEmitter).removeAllListeners(SOCKET_ERROR);
+        (this as EventEmitter).removeAllListeners(SOCKET_CONNECT);
     }
     /**
      * Will close the socket connection with a error code and reason.
@@ -201,5 +317,120 @@ export class WebSocketProvider extends SocketProvider {
                     });
             });
         });
+    }
+    /**
+     * Subscribes to a given subscriptionType
+     * @param subscribeMethod JSON-RPC method name to use for subscription
+     * @param subscriptionMethod Subscription namespace
+     * @param parameters Additional parameters to subscribe call
+     * @returns The subscriptionId of an error
+     */
+    public subscribeAsync(subscribeMethod: string, subscriptionMethod: string, parameters: any[]): Promise<any> {
+        parameters.unshift(subscriptionMethod);
+
+        return this.sendAsync(subscribeMethod, parameters)
+            .then(subscriptionId => {
+                this.subscriptions[subscriptionId] = {
+                    id: subscriptionId,
+                    subscribeMethod,
+                    parameters,
+                };
+
+                return subscriptionId;
+            })
+            .catch(error => {
+                throw new Error(`Provider error: ${error}`);
+            });
+    }
+    /**
+     * Unsubscribes the subscription by his id
+     * @param subscriptionId subscription identifier corresponding to the subscription to cancel
+     * @param unsubscribeMethod The JSON-RPC subscription method name
+     *
+     * @returns either a boolean of whether the subscription was cancelled, or an error
+     */
+    public unsubscribeAsync(subscriptionId: string, unsubscribeMethod: string): Promise<boolean|Error> {
+        if (this._hasSubscription(subscriptionId)) {
+            return this.sendAsync(unsubscribeMethod, [subscriptionId]).then(response => {
+                if (response) {
+                    (this as EventEmitter).removeAllListeners(this._getSubscriptionEvent(subscriptionId));
+
+                    delete this.subscriptions[subscriptionId];
+                }
+
+                return response;
+            });
+        }
+
+        return Promise.reject(new Error(`Provider error: Subscription with ID ${subscriptionId} does not exist.`));
+    }
+    /**
+     * Clears all subscriptions and listeners
+     * @param unsubscribeMethod JSON-RPC unsubscribe method
+     * @returns true if clearing subscription succeeds, otherwise an error
+     */
+    public async clearSubscriptionsAsync(unsubscribeMethod: string): Promise<boolean|Error> {
+        const unsubscribePromises: Array<Promise<any>> = [];
+
+        Object.keys(this.subscriptions).forEach(key => {
+            (this as EventEmitter).removeAllListeners(key);
+            unsubscribePromises.push(this.unsubscribeAsync(this.subscriptions[key].id, unsubscribeMethod));
+        });
+
+        return Promise.all(unsubscribePromises).then(results => {
+            if (results.includes(false)) {
+                throw new Error(`Could not unsubscribe all subscriptions: ${JSON.stringify(results)}`);
+            }
+
+            return true;
+        });
+    }
+    /**
+     * Checks if the given subscription id exists
+     * @param subscriptionId subscription ID to check existence for
+     * @returns whether or not the subscription exists
+     */
+    private _hasSubscription(subscriptionId: string): boolean {
+        return typeof this._getSubscriptionEvent(subscriptionId) !== 'undefined';
+    }
+    /**
+     * Returns the event the subscription is listening for.
+     * @param subscriptionId subscription ID
+     * @returns subscription event name (e.g. "heartbeat")
+     */
+    private _getSubscriptionEvent(subscriptionId: string): string|undefined {
+        if (this.subscriptions[subscriptionId]) {
+            return subscriptionId;
+        }
+
+        let event: string|undefined;
+        for (const key in this.subscriptions) {
+            if (this.subscriptions[key].id === subscriptionId) {
+                event = key;
+            }
+        }
+
+        return event;
+    }
+    /**
+     * Creates a valid json payload object
+     * @param method JSON-RPC method to call
+     * @param params parameters to supply in method call
+     * @returns JSON-RPC payload
+     */
+    private _toPayload(method: string, params: any[]): any {
+        if (!method) {
+            throw new Error(`JSONRPC method should be specified for params: "${JSON.stringify(params)}"!`);
+        }
+
+        const id = this._messageId;
+        this._messageId++;
+
+        return {
+            jsonrpc: '2.0',
+            id,
+            method,
+            params: params || [],
+        };
     }
 }
